add fillOrderForm page 1 tests
